+++ src/components/AppBar/AppBar.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import ModeSelect from '~/components/ModeSelect'
+import ModeSelect from '~/components/ModeSelect/ModeSelect'
 import Box from '@mui/material/Box'
 import AppsIcon from '@mui/icons-material/Apps'
 import { ReactComponent as TrelloIcon } from '~/assets/mdi--trello.svg'
