+++ src/pages/Boards/BoardContent/BoardContent.jsx
@@ -1,0 +1,17 @@
+import Box from '@mui/material/Box'
+import ListColumns from './ListColumns/ListColumns'
+
+function BoardContent() {
+  return (
+    <Box sx={{
+      bgcolor: (theme) => (theme.palette.mode === 'dark' ? '#323b53' : '#9c88ff'),
+      width: '100%',
+      height: (theme) => theme.trello.boardContentHeight,
+      p: '10px 0'
+    }}>
+      <ListColumns />
+    </Box>
+  )
+}
+
+export default BoardContent
