+++ src/pages/Boards/_id.jsx
@@ -1,7 +1,7 @@
 import Container from '@mui/material/Container'
-import BoardBar from './BoardBar'
-import AppBar from '~/components/AppBar'
-import BoardContent from './BoardContent'
+import AppBar from '~/components/AppBar/AppBar'
+import BoardBar from './BoardBar/BoardBar'
+import BoardContent from './BoardContent/BoardContent'
 
 function Board() {
   return (
