+++ src/theme.js
@@ -3,12 +3,16 @@ import { experimental_extendTheme as extendTheme } from '@mui/material/styles'
 const APP_BAR_HEIGHT = '58px'
 const BOARD_BAR_HEIGHT = '60px'
 const BOARD_CONTENT_HEIGHT = `calc(100vh - (${APP_BAR_HEIGHT} + ${BOARD_BAR_HEIGHT}))`
+const COULUMN_HEADER_HEIGHT = '50px'
+const COULUMN_FOOTER_HEIGHT = '50px'
 // Create a theme instance.
 const theme = extendTheme({
   trello: {
     appBarHeight: APP_BAR_HEIGHT,
     boardBarHeight: BOARD_BAR_HEIGHT,
-    boardContentHeight: BOARD_CONTENT_HEIGHT
+    boardContentHeight: BOARD_CONTENT_HEIGHT,
+    columnHeaderHeight: COULUMN_HEADER_HEIGHT,
+    columnFooterHeight: COULUMN_FOOTER_HEIGHT
   },
   colorSchemes: {
     light: {},
